Move functionality to existing ChooseAssayTypePage
